requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
streamlit==1.28.1
requests==2.31.0
python-dotenv==1.0.0
pandas>=2.2.0
plotly==5.18.0

